backend/requirements.txt
pyaudio==0.2.14
deepgram-sdk==3.4.0
google-generativeai==0.3.2
elevenlabs==0.2.27
pygame==2.5.2
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
psutil==5.9.6
openai==1.12.0
pydub==0.25.1
av==12.0.0
pyautogui==0.9.54
selenium==4.16.0
webdriver-manager==4.0.1
pillow==10.2.0
keyboard==0.13.5
mouse==0.7.1